requirements.txt
torch>=1.9.0
numpy>=1.19.0
pandas>=1.2.0
scipy>=1.5.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
tqdm>=4.50.0
pyyaml>=5.4.0
torchcde>=0.2.5
torchdiffeq>=0.2.3
